updates to products server

+++ src/main/java/com/motocart/products_microservice/repository/entity/ProductsEntity.java
@@ -17,5 +17,5 @@ public class ProductsEntity {
     private String productCode;
     private String firmName;
     private String description;
-    private double price;
+    private Long price;
 }
